Revert "remove docs/_static folder."
This reverts commit e5351cfe957cfacab60dded18bc825aa5a4a685b.

As mentioned on the issue tracker, [0] some of the CSS from docs/_static was
still required to make the layout of the Sphinx generated output have the
correct layout. I reverted the last commit, which removed the whole
docs/_static directory, and then removed all the JavaScript and the unneeded
CSS. Hopefully, this is okay with Debian. :)

[0]: https://github.com/isislovecruft/python-gnupg/pull/27#issuecomment-29712108