Refactor _util._is_file to handle detached sigs and data on stdin.
When gnupg.GPG.verify_file() was changed recently to take the filename
of the signature file as an argument to GnuPG, and then take the data
file to be verified on stdin, the parser in _parsers._check_options()
would call _util._is_file(), which would return False and log an error
that "'-' is not a file!". This fixes that issue by catching OSError and
treating it differently.

Additionally, I renamed the _util._is_file() parameter
`input`→`filename` to avoid overriding a builtin method.

I also replaced the `assert` lines with explicit checks, because
`assert`s are stripped out when the Python interpreter is run with -OO.

 * FIXES an issue with verification of detached signatures whose
   datafiles are read from stdin.