Workaround possible upstream GnuPG bug sign fail when passphrase=''.
If passed `--passphrase-fd 0` and piped the passphrase `''` (an empty
string), GnuPG will truncate the message being signed. The truncation
appears to only affect messages which contain newlines; for those
messages, the first line (up unto and including the newline character)
will be truncated and thus missing from the output signed message or
signature, causing verification for signatures on the original
non-truncated message to fail.

 * FIXES #82: https://github.com/isislovecruft/python-gnupg/issues/82